refactor

+++ modules/marketdata-processor/src/main/java/com/example/marketdata/stats/metric/ILatencyMetric.java
@@ -3,5 +3,5 @@ package com.example.marketdata.stats.metric;
 public interface ILatencyMetric {
     void record(long latency);
     LatencyValues snapshotAndReset();
-    record LatencyValues(double avg, double max) {}
+    record LatencyValues(double avg, long max) {}
 }
